ipl_prediction/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
bs4==0.0.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
Flask==3.1.1
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
keras==3.10.0
keras-models==0.0.7
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
murmurhash==1.0.12
namex==0.0.9
numpy==2.2.5
opencv-python==4.11.0.86
optree==0.15.0
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
preshed==3.0.9
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.8.0
shellingham==1.5.4
six==1.17.0
sklearn-preprocessing==0.1.0
smart-open==7.1.0
soupsieve==2.7
spacy==3.8.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.3.6
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
